+++ src/data/diffAgainstStore.ts
@@ -1,5 +1,6 @@
 import isArray = require('lodash.isarray');
 import isNull = require('lodash.isnull');
+import isUndefined = require('lodash.isundefined');
 import has = require('lodash.has');
 import assign = require('lodash.assign');
 
@@ -15,6 +16,10 @@ import {
   isJsonValue,
   isIdValue,
 } from './store';
+
+import {
+  StoreFetchMiddleware,
+} from './fetchMiddleware';
 
 import {
   SelectionSetWithRoot,
@@ -48,6 +53,7 @@ export interface DiffResult {
 export interface StoreContext {
   store: NormalizedCache;
   fragmentMap: FragmentMap;
+  fetchMiddleware?: StoreFetchMiddleware;
 }
 
 export function diffQueryAgainstStore({
@@ -251,7 +257,17 @@ function diffFieldAgainstStore({
   const storeObj = context.store[rootId] || {};
   const storeFieldKey = storeKeyNameFromField(field, variables);
 
-  if (! has(storeObj, storeFieldKey)) {
+  let storeValue, fieldMissing;
+  // Give the transformer a chance to yield a rewritten result.
+  if (context.fetchMiddleware) {
+    storeValue = context.fetchMiddleware(field, variables, context.store, () => storeObj[storeFieldKey]);
+    fieldMissing = isUndefined(storeValue);
+  } else {
+    storeValue = storeObj[storeFieldKey];
+    fieldMissing = !has(storeObj, storeFieldKey);
+  }
+
+  if (fieldMissing) {
     if (throwOnMissingField && included) {
       throw new Error(`Can't find field ${storeFieldKey} on object ${JSON.stringify(storeObj)}.
 Perhaps you want to use the \`returnPartialData\` option?`);
@@ -261,8 +277,6 @@ Perhaps you want to use the \`returnPartialData\` option?`);
       isMissing: 'true',
     };
   }
-
-  const storeValue = storeObj[storeFieldKey];
 
   // Handle all scalar types here
   if (! field.selectionSet) {